refactor: User, Auth 책임 분리 #99
