+++ BE/ytt/src/main/java/com/example/ytt/domain/vendingmachine/controller/VendingMachineManageController.java
@@ -1,7 +1,7 @@
 package com.example.ytt.domain.vendingmachine.controller;
 
 import com.example.ytt.domain.inventory.dto.InboundReqDto;
-import com.example.ytt.domain.user.auth.security.CustomUserDetails;
+import com.example.ytt.domain.auth.security.CustomUserDetails;
 import com.example.ytt.domain.user.dto.Role;
 import com.example.ytt.domain.vendingmachine.domain.MachineState;
 import com.example.ytt.domain.vendingmachine.dto.VendingMachineDetailDto;
